Changes due to model-library refactor
