+++ services/catalog/src/simcore_service_catalog/models/schemas/dag.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional
 
-from models_library.projects import Node
+from models_library.projects_nodes import Node
 
 from ..domain.dag import DAGBase, DAGData
