+++ services/director-v2/src/simcore_service_director_v2/models/domains/comp_pipelines.py
@@ -1,6 +1,7 @@
 from typing import Dict, List
 
-from models_library.projects import ProjectID, RunningState
+from models_library.projects import ProjectID
+from models_library.projects_state import RunningState
 from pydantic import BaseModel, validator
