+++ services/director-v2/src/simcore_service_director_v2/models/schemas/meta.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional
 
 from pydantic import BaseModel, Field, constr
 
-from models_library.constants import VERSION_RE
+from models_library.basic_regex import VERSION_RE
 
 VersionStr = constr(regex=VERSION_RE)
